Eglot: rework semtok feature again
This version was tested with the clangd, rust-analyzer and lean servers.
Each server is slightly different.

The main change is to recognize that that eglot--semtok-request is a
promise to get some data for BEG to END and a promise to flush it.  The
eglot--semtok-inflight variable is overhauled so that it more precisely
encodes this information.

The lean server insists on many workspace/semanticTokens/refresh waiting
for a textDocument/semanticTokens/full request from our part, but we
don't give in.  The spec doesn't actually say what types of requests the
clients should issue.  It eventualy gives up.  The only thing we do on
this request is to flush everything we know.  After this flush, the
font-lock machinery will know what to do.

* lisp/progmodes/eglot.el (eglot--async-request): Return id.
(eglot--semtok-inflight): Redesign variable.
(eglot-handle-request): Just flush.
(eglot--semtok-request): Rewrite.
(eglot--semtok-font-lock): Tweak.