Improve robustness of 'package-upgrade'
* lisp/package/package-install.el (package-install): Raise an
error if the package was not installed.  This prevents
'package-upgrade' from proceeding to remove the old package.
(package-upgrade): Install new version of package via package
descriptor, so that we can first install the new version of a
package before uninstalling the old version, thus reducing the
chance of a package upgrade resulting in an inconsistent state.
* test/lisp/emacs-lisp/package-tests.el
(package-test-update-listing): Fix test by refreshing the
package buffer to indicate the actual state.